make ExpectClusterQueueWeightedShareMetric accept float64 as value

+++ test/util/util.go
@@ -728,7 +728,7 @@ func ExpectClusterQueueStatusMetric(cq *kueue.ClusterQueue, status metrics.Clust
 	}
 }
 
-func ExpectClusterQueueWeightedShareMetric(cq *kueue.ClusterQueue, value int64) {
+func ExpectClusterQueueWeightedShareMetric(cq *kueue.ClusterQueue, value float64) {
 	metric := metrics.ClusterQueueWeightedShare.WithLabelValues(cq.Name, string(cq.Spec.Cohort))
 	gomega.EventuallyWithOffset(1, func(g gomega.Gomega) {
 		v, err := testutil.GetGaugeMetricValue(metric)
